% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceEnvelopes.R
\encoding{UTF-8}
\name{envelope}
\alias{envelope}
\alias{trajectoryEnvelopeVariability}
\alias{stateEnvelopeVariability}
\alias{compareToStateEnvelope}
\alias{compareToTrajectoryEnvelope}
\title{Reference envelopes}
\usage{
trajectoryEnvelopeVariability(
  d,
  sites,
  surveys = NULL,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  ...
)

stateEnvelopeVariability(d, nboot.ci = NULL, alpha.ci = 0.05)

compareToTrajectoryEnvelope(
  d,
  sites,
  envelope,
  surveys = NULL,
  m = 1.5,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  ...
)

compareToStateEnvelope(
  d,
  envelope,
  m = 1.5,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  ...
)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states (see details).}

\item{sites}{A vector indicating the site corresponding to each ecosystem state.}

\item{surveys}{A vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{nboot.ci}{Number of bootstrap samples for confidence intervals. If nboot.ci = NULL then confidence intervals are not estimated.}

\item{alpha.ci}{Error in confidence intervals.}

\item{...}{Additional parameters for function \code{\link{trajectoryDistances}}}

\item{envelope}{A vector indicating the set of sites that conform the reference envelope (other sites will be compared to the envelope)}

\item{m}{Fuzziness exponent for quality value assessment}
}
\description{
Reference envelopes
}
\seealso{
\code{\link{trajectorymetrics}}
}
\author{
Miquel De \enc{Cáceres}{Caceres}, CREAF

Anthony Sturbois, Vivarmor nature, Réserve Naturelle nationale de la Baie de Saint-Brieuc
}
