% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceEnvelopes.R
\encoding{UTF-8}
\name{envelopes}
\alias{envelopes}
\alias{envelopeVariability}
\alias{segmentDistances}
\alias{trajectoryDistances}
\alias{trajectoryLengths}
\alias{trajectoryLengths2D}
\alias{trajectoryAngles}
\alias{trajectoryAngles2D}
\alias{trajectoryProjection}
\alias{trajectoryConvergence}
\alias{trajectoryDirectionality}
\title{Reference envelopes}
\usage{
envelopeVariability(
  d,
  sites,
  surveys = NULL,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  ...
)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states (see details).}

\item{sites}{A vector indicating the site corresponding to each ecosystem state.}

\item{surveys}{A vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{...}{Additional parameters for function \code{\link{trajectoryDistances}}}
}
\description{
Reference envelopes
}
\seealso{
\code{\link{trajectorymetrics}}
}
\author{
Miquel De \enc{Cáceres}{Caceres}, CREAF

Anthony Sturbois, Vivarmor nature, Réserve Naturelle nationale de la Baie de Saint-Brieuc
}
