% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMA.R
\name{RTMA}
\alias{RTMA}
\title{Relative Trajectory Movement Assessment (RTMA)}
\usage{
RTMA(x, alpha = 0.05, nperm = 999, full.output = TRUE, add = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{trajectories}}.}

\item{alpha}{The alpha level for the tests performed in RTMA. Defaults to \code{0.05}.}

\item{nperm}{Passed to function \code{\link{trajectoryCorrespondence}}. The number of permutations to be used in the dynamic correspondence test. Defaults to \code{999}.}

\item{full.output}{Flag to indicate that the full output of tests should be computed. Defaults to \code{TRUE}. Setting to FALSE will improve computation speed but yield incomplete outputs (see details).}

\item{add}{Passed to function \code{\link{trajectoryConvergence}}. Flag to indicate that constant values should be added (local transformation) to correct triplets of distance values that do not fulfill the triangle inequality.}
}
\value{
Function \code{RTMA} returns a object of class \code{\link{list}} containing:
\itemize{
\item{\code{ScenarioAttribution}, a matrix containing the relative movements scenario attributed to each pair of trajectories.}
\item{\code{DetailedTests}, containing the details of the different tests performed by RTMA with:
\itemize{
\item{\code{SymmetricConvergence} contains the results of the symmetric convergence test.}
\item{\code{AsymmetricConvergence} contains the results of the two asymmetric convergence tests.}
\item{\code{DynamicCorrespondence} contains the results of the the dynamic correspondence tests (potentially partial if \code{full.out = FALSE}).}
}
}
}
}
\description{
Relative Trajectory Movement Assessment (RTMA) is a method allowing testing and qualifying of the relative movements of ecological trajectories (e.g. "convergence", "parallel" etc., see details) as described in Djeghri et al. (in prep).
}
\details{
Function \code{RTMA} attributes a "scenario" of relative movement to pairs of ecological trajectories A and B by combining four tests:
\itemize{
\item{Three convergence tests performed through internal callings of function \code{\link{trajectoryConvergence}}:
\itemize{
\item{The symmetric convergence test between trajectories A and B.}
\item{The asymmetric convergence test assessing if trajectory A approaches trajectory B.}
\item{The asymmetric convergence test assessing if trajectory B approaches trajectory A.}
}
}
\item{One dynamic correspondence test performed through internal callings to function \code{\link{trajectoryCorrespondence}}.}
}
To account for multiple testing, \code{RTMA} performs internally a \enc{Šidák}{Sidak} (1967) correction on the alpha level provided in parameter \code{alpha}.

The results of the four tests (p-values and sign of statistic) are used to assign to each trajectory pair a "scenario" describing their relative movements. RTMA recognizes a total of 10 scenarios, some existing in "weak" variations:
\itemize{
\item{\code{Convergence} scenario: The two trajectories converge, exists in a weak version.}
\item{\code{Divergence} scenario: The two trajectories diverge, exists in a weak version.}
\item{\code{Approaching} scenario: One trajectory approaches the other (likely relatively immobile), exists in a weak version.}
\item{\code{Departing} scenario: One trajectory moves away from the other (likely relatively immobile), exists in a weak version.}
\item{\code{Pursuit} scenario: The two trajectories follow each other.}
\item{\code{Catchup} scenario: As \code{Pursuit} but the following trajectory moves faster.}
\item{\code{Escape} scenario: As \code{Pursuit} but the leading trajectory is faster.}
\item{\code{Parallel} scenario: The two trajectories travel side by side with broadly similar movements.}
\item{\code{Antiparallel} scenario: As \code{Parallel} but the two trajectories travel in opposite directions.}
\item{\code{Neutral} scenario:  The two trajectories have no particular movements relative to each other (effectively the Null Hypothesis).}
}
In rare cases, unlikely scenarios (labelled \code{Other}) may occur. These involve contradictory patterns hard to interpret.

LIMITATIONS: RTMA has some limitations, in particular it uses trend tests not well suited to study trajectories pairs with changing relative movements (e.g. if two trajectories cross each other, they are first converging then diverging).
We advise users to not only rely on RTMA but to also visualize trajectories using function \code{\link{trajectoryPCoA}} for ecological interpretations. See Djeghri et al. (in prep) for more details.

COMPUTATION TIME: The dynamic correspondence tests performed in RTMA are computationally costly permutation tests only used when all three convergence tests are non-significant.
Function \code{RTMA} performs by default all tests but it is possible to only perform the tests useful for RTMA by setting \code{full.output = FALSE}.
This will reduce computation time but the details of the output of RTMA will not contain the information on all possible dynamic correspondence tests, only on relevant ones.

PLOTTING: Function \code{\link{trajectoryConvergencePlot}} provides options to plot the results of RTMA.
}
\references{
Djeghri et al. (in preparation) Uncovering the relative movements of ecological trajectories.

\enc{Šidák}{Sidak}, Z. (1967) Rectangular confidence regions for the means of multivariate normal distributions. Journal of the American Statistical Association 62:648-633.
}
\seealso{
\code{\link{trajectoryConvergence}}, \code{\link{trajectoryCorrespondence}}, \code{\link{trajectoryConvergencePlot}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
