% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentDistances.R
\name{segmentDistances}
\alias{segmentDistances}
\title{Distances between segments}
\usage{
segmentDistances(x, distance.type = "directed-segment", add = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{trajectories}}.}

\item{distance.type}{The type of distance index to be calculated (Besse et al. 2016):
\itemize{
\item{\code{Hausdorff}: Hausdorff distance between two segments.}
\item{\code{directed-segment}: Directed segment distance (default).}
\item{\code{PPA}: Perpendicular-parallel-angle distance.}
}}

\item{add}{}
}
\value{
A list with the following elements:
\itemize{
\item{\code{Dseg}: Distance matrix between segments.}
\item{\code{Dini}: Distance matrix between initial points of segments.}
\item{\code{Dfin}: Distance matrix between final points of segments.}
\item{\code{Dinifin}: Distance matrix between initial points of one segment and the final point of the other.}
\item{\code{Dfinini}: Distance matrix between final points of one segment and the initial point of the other.}
}
}
\description{
Calculates the distance between pairs of directed segments of input trajectories.
}
\details{
Details of calculations are given in De \enc{Cáceres}{Caceres} et al (2019).
The input distance matrix \code{d} should ideally be metric. That is, all subsets of distance triplets should fulfill the triangle inequality (see utility function \code{\link{is.metric}}).
All ETA functions that require metricity include a parameter '\code{add}', which by default is TRUE, meaning that whenever the triangle inequality is broken the minimum constant required to fulfill it is added to the three distances.
If such local (an hence, inconsistent across triplets) corrections are not desired, users should find another way modify \code{d} to achieve metricity, such as PCoA, metric MDS or non-metric MDS (see vignette 'Introduction to Ecological Trajectory Analysis').
If parameter '\code{add}' is set to FALSE and problems of triangle inequality exist, ETA functions may provide missing values in some cases where they should not.
}
\references{
Besse, P., Guillouet, B., Loubes, J.-M. & François, R. (2016). Review and perspective for distance based trajectory clustering. IEEE Trans. Intell. Transp. Syst., 17, 3306–3317.

De \enc{Cáceres}{Caceres} M, Coll L, Legendre P, Allen RB, Wiser SK, Fortin MJ, Condit R & Hubbell S. (2019).
Trajectory analysis in community ecology. Ecological Monographs 89, e01350.
}
