% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryConvergencePlot.R
\encoding{UTF-8}
\name{trajectoryCyclicalPlots}
\alias{trajectoryCyclicalPlots}
\alias{trajectoryConvergencePlot}
\title{Summary plot for trajectory convergence and divergence}
\usage{
trajectoryConvergencePlot(
  x,
  type = "pairwise.asymmetric",
  alpha.filter = NULL,
  traj.colors = "grey",
  traj.names = NULL,
  traj.names.colors = "black",
  ...,
  radius = 1,
  conv.color = "red",
  div.color = "blue",
  half.arrows.size = 1,
  tau.links.transp = 0.3,
  top = "between",
  pointy = FALSE,
  CS = NULL,
  CSinfConf = NULL,
  CSsupConf = NULL,
  coeffCS = 0.5,
  lwd.arrows.CS = 3
)
}
\arguments{
\item{x}{An object of class \code{\link{trajectories}}.}

\item{type}{A string indicating the convergence test, either "pairwise.asymmetric", "pairwise.symmetric" or "both" (see \code{\link{trajectoryConvergence}}).}

\item{alpha.filter}{The minimum p-value for a link to be drawn (see \code{\link{trajectoryConvergence}}). Defaults to NULL (all links drawn).}

\item{traj.colors}{The colors for the trajectories (circles). Defaults to "grey".}

\item{traj.names}{The names of trajectories. Defaults to the names provided in \code{x}.}

\item{traj.names.colors}{The color of the names of trajectories on the circles. Defaults to "black".}

\item{...}{Additional parameters passed to \code{\link{polygon}} to personalize the circles representing trajectories.}

\item{radius}{The radius of the circles representing trajectories. Defaults to 1.}

\item{conv.color}{The color used to mark convergent trajectories. Defaults to "red".}

\item{div.color}{The color used to mark divergent trajectories. Defaults to "blue".}

\item{half.arrows.size}{A multiplication coefficient for the size of the arrow heads when representing asymmetric tests results. Defaults to 1.}

\item{tau.links.transp}{The transparency of the links representing the tau statistic of the Mann.Kendall test (see \code{\link{trajectoryConvergence}}).}

\item{top}{A string indicating whether the top of the plotting area should contain a circle representing a trajectory ("top") or should be in between two circles ("between"). Defaults to "between".}

\item{pointy}{Boolean. Should the circles representing trajectories be made pointy? Useful in the context of CETA to represent fixed date trajectories (see \code{\link{trajectoryCyclical}}).}

\item{CS}{Experimental. Allows to add arrows representing cyclical shifts in the context of CETA.}

\item{CSinfConf}{Experimental. Lower confidence intervals for cyclical shifts.}

\item{CSsupConf}{Experimental. Upper confidence intervals for cyclical shifts.}

\item{coeffCS}{Experimental. A multiplication coefficient for the arrows representing cyclical shifts.}

\item{lwd.arrows.CS}{Experimental. The width of the arrows representing cyclical shifts.}
}
\description{
Provides plots to represent trajectory convergence and divergence tests performed by the function \code{\link{trajectoryConvergence}}.
}
\details{
The function \code{trajectoryConvergencePlot} provides ways to visualize pairwise convergence and divergence between trajectories using calls to function \code{\link{trajectoryConvergence}} which performs the tests.
In the plots, trajectories are represented by circles. The convergence or divergence between pairs of trajectories are represented by links. If convergence tests are symmetric, the links are simple. If the convergence tests are asymmetric, the links are displayed as half arrows pointing from the trajectory converging or diverging towards the trajectory being approached or diverged from.
The width and color hue of the links are proportional to the tau statistic of the Mann.Kendall test performed by the \code{\link{trajectoryConvergence}} function.
The function \code{trajectoryConvergencePlot} also offers the possibility to plot both tests at the same time.

The function finally offers some possibilities more relevant in the context of cyclical ecological trajectory analysis (CETA) such as the possibility to make the circles representing trajectories "pointy" so that they communicate a cyclical organization. This is useful if fixed dates trajectories are being studied (see \code{\link{extractFixedDateTrajectories}}).
The function also allows to display "cyclical shifts" but values have to be provided by the user as they can be obtained in a variety of ways summarizing the outputs of \code{\link{cycleShifts}}.
}
\examples{
data("avoca")
avoca_D_man <- vegclust::vegdiststruct(avoca_strat, 
                                       method="manhattan", 
                                       transform = function(x){log(x+1)})
years <- c(1971, 1974, 1978, 1983, 1987, 1993, 1999, 2004, 2009)
avoca_times <- years[avoca_surveys]
avoca_x <- defineTrajectories(d = avoca_D_man,  
                              sites = avoca_sites, 
                              times = avoca_times)

#Raw output with asymmetric convergence test (default)
trajectoryConvergencePlot(avoca_x)

#More refined output with both type of tests and only plotting significant test results (p-value < 0.05)
trajectoryConvergencePlot(avoca_x,type="both",alpha.filter=0.05)

#Much more refined output with nicer colors bigger half arrows and personalized trajectory names
trajectoryConvergencePlot(avoca_x,type="both",alpha.filter=0.05,
                          #adjusting the attributes of the links between trajectories
                          half.arrows.size = 1.5,conv.color = "orangered",div.color="dodgerblue",
                          #controling the size of circles representing trajectories (links are enlarged to match it)
                          radius=1.2, 
                          #customizing the border of the circles representing trajectories
                          traj.colors = "black",border="white",lwd=2,
                          #giving trajectories new names and adjusting the names color
                          traj.names=LETTERS[1:8],traj.names.colors="white")


}
\seealso{
\code{\link{trajectoryConvergence}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
