% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryCyclicalPlots.R
\encoding{UTF-8}
\name{trajectoryplots}
\alias{trajectoryplots}
\alias{cyclePCoA}
\alias{fixedDateTrajectoryPCoA}
\alias{CETA}
\alias{plotting}
\alias{functions:}
\alias{\itemize{}
\alias{\item{Function}
\alias{\code{cyclePCoA}}
\alias{performs}
\alias{principal}
\alias{coordinates}
\alias{analysis}
\alias{(\code{\link{cmdscale}})}
\alias{and}
\alias{draws}
\alias{cycles}
\alias{in}
\alias{the}
\alias{ordination}
\alias{scatterplot.}
\alias{Sister}
\alias{function}
\alias{of}
\alias{\code{trajectoryPCoA}}
\alias{adapted}
\alias{to}
\alias{cycles.}}
\alias{}}
\title{Cyclical trajectory plots}
\usage{
cyclePCoA(
  d,
  cycles,
  times,
  IntExt,
  selection = NULL,
  cycle.colors = NULL,
  axes = c(1, 2),
  time.labels = FALSE,
  ...
)

fixedDateTrajectoryPCoA()
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states.}

\item{cycles}{A vector indicating the cycles corresponding to each ecosystem state.}

\item{times}{A vector indicating the times corresponding to each ecosystem state (equivalent to "surveys" in other ETA function but more time-explicit).}

\item{IntExt}{A vector indicating whether each ecosystem state is "internal" or "external".}

\item{selection}{A character vector of cycles, a numeric vector of cycle indices or logical vector of the same length as \code{cycles}, indicating a subset of cycle trajectories to be selected.}

\item{cycle.colors}{A vector of colors (one per cycle). If \code{selection != NULL} the length of the color vector should be equal to the number of cycles selected.}

\item{axes}{The pair of principal coordinates to be plotted.}

\item{time.labels}{A boolean flag to indicate whether times should be plotted as text next to arrow endpoints}

\item{...}{Additional parameters for function \code{\link{arrows}}.}
}
\value{
Function \code{cyclePCoA} returns a list similar to the calling \code{\link{cmdscale}}.
}
\description{
Set of plotting functions for Cyclical Ecological Trajectory Analysis:
}
\details{
\itemize{
\item{Function \code{cyclePCoA} performs principal coordinates analysis (\code{\link{cmdscale}}) and draws trajectories in the ordination scatterplot.}
\item{Function \code{fixedDateTrajectoryPCoA} performs principal coordinates analysis (\code{\link{cmdscale}}) and draws trajectories in the ordination scatterplot.}
}

Additional details...
}
\seealso{
\code{\link{trajectoryCyclical}}}, \code{\link{cmdscale}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
