% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectorySmoothing.R
\name{trajectorySmoothing}
\alias{trajectorySmoothing}
\title{Trajectory smoothing}
\usage{
trajectorySmoothing(
  d,
  sites,
  surveys = NULL,
  survey_times = NULL,
  kernel_scale = 1,
  fixed_endpoints = TRUE
)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states.}

\item{sites}{A vector indicating the site corresponding to each ecosystem state.}

\item{surveys}{A vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{survey_times}{A vector indicating the survey time for all surveys (if \code{NULL}, time between consecutive surveys is considered to be one)}

\item{kernel_scale}{Scale of the Gaussian kernel, related to survey times}

\item{fixed_endpoints}{A logical flag to force keeping the location of trajectory endpoints unmodified}
}
\value{
An object of class \code{\link{dist}}
}
\description{
Trajectory smoothing using a Gaussian kernel
}
