% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleShiftArrows.R
\encoding{UTF-8}
\name{cycleShiftArrows}
\alias{cycleShiftArrows}
\title{Displaying cycle shifts}
\usage{
cycleShiftArrows(
  cycle.shifts,
  radius = 1,
  top = "between",
  cycle.shifts.inf.conf = NULL,
  cycle.shifts.sup.conf = NULL,
  arrows.length.mult = "auto",
  arrows.lwd = 2
)
}
\arguments{
\item{cycle.shifts}{Cyclical shifts computed for each fixed date trajectory plotted by \code{trajectoryConvergencePlot}.}

\item{radius}{The radius of the circles representing trajectories. Defaults to 1.}

\item{top}{A string indicating if the top of the plotting area should contain a circle representing a trajectory ("circle"), or should be in between two circles ("between"). Defaults to "between".}

\item{cycle.shifts.inf.conf}{Lower confidence intervals for cyclical shifts.}

\item{cycle.shifts.sup.conf}{Upper confidence intervals for cyclical shifts.}

\item{arrows.length.mult}{A multiplication coefficient for the arrows representing cyclical shifts. Attempts an automatic adjustment by default (dividing by max(cycle.shifts)).}

\item{arrows.lwd}{Line width of the arrows. Defaults to 2.}
}
\description{
Adds arrows representing cyclical shifts (advances/delays) into convergence/divergence plots created using function \code{\link{trajectoryConvergencePlot}}.
}
\details{
This function is meant to be used in conjunction with \code{\link{trajectoryConvergencePlot}}, to study fixed-date trajectories convergence/divergence patterns:
\itemize{
\item{First, setting \code{pointy = TRUE}, in the call to \code{\link{trajectoryConvergencePlot}}, when the studied trajectories are the outputs of \code{\link{extractFixedDateTrajectories}} allows to suggest cyclicity.}
\item{Second, function \code{cycleShiftArrows} allows to represent cyclical shifts by adding arrows to the circles representing trajectories. Clockwise arrows will represent advances, anticlockwise arrows will represent delays. Arrows length is proportional to the cyclical shift provided.
Relevant measures of cyclical shifts have to be computed by the user. A variety of methods may be employed but the outputs of \code{\link{cycleShifts}} cannot be used immediately as they do not directly correspond to the fixed-date trajectories. An example of how to compute relevant
measures of cyclical shifts is provided in \code{\link{cycleShifts}} and in the CETA vignette.
The arguments \code{radius} and \code{top} in \code{cycleShiftArrows} must match those in the corresponding call to \code{\link{trajectoryConvergencePlot}} for proper display.}
}
}
\examples{
data("avoca")
avoca_D_man <- vegclust::vegdiststruct(avoca_strat, 
                                       method ="manhattan", 
                                       transform = function(x){log(x+1)})
years <- c(1971, 1974, 1978, 1983, 1987, 1993, 1999, 2004, 2009)
avoca_times <- years[avoca_surveys]
avoca_x <- defineTrajectories(d = avoca_D_man,  
                              sites = avoca_sites, 
                              times = avoca_times)


#CETA-specific example of convergence/divergence plot with made-up cyclical shifts:
trajectoryConvergencePlot(avoca_x,
                          type = "both",
                          alpha.filter = 0.05,
                          half.arrows.size = 2,
                          pointy = TRUE)

#Made-up cyclical shifts (matching the number of trajectories in avoca_x)
CS <- c(2.5, 1.5, 0.5, 0.5, 3, 1, 1, 2)
cycleShiftArrows(CS,
                  cycle.shifts.inf.conf = CS - 0.2,
                  cycle.shifts.sup.conf = CS + 0.2)
}
\seealso{
\code{\link{trajectoryConvergencePlot}},\code{\link{cycleShifts}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
