% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceEnvelopes.R
\encoding{UTF-8}
\name{envelope}
\alias{envelope}
\alias{trajectoryEnvelopeVariability}
\alias{stateEnvelopeVariability}
\alias{compareToStateEnvelope}
\alias{compareToTrajectoryEnvelope}
\title{Reference envelopes}
\usage{
trajectoryEnvelopeVariability(
  d,
  sites,
  surveys = NULL,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  ...
)

stateEnvelopeVariability(d, nboot.ci = NULL, alpha.ci = 0.05)

compareToTrajectoryEnvelope(
  d,
  sites,
  envelope,
  surveys = NULL,
  m = 1.5,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  distances_to_envelope = FALSE,
  ...
)

compareToStateEnvelope(
  d,
  envelope,
  m = 1.5,
  nboot.ci = NULL,
  alpha.ci = 0.05,
  distances_to_envelope = FALSE,
  ...
)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states (see details).}

\item{sites}{A vector indicating the site corresponding to each ecosystem state.}

\item{surveys}{A vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{nboot.ci}{Number of bootstrap samples for confidence intervals. If nboot.ci = NULL then confidence intervals are not estimated.}

\item{alpha.ci}{Error in confidence intervals.}

\item{...}{Additional parameters for function \code{\link{trajectoryDistances}}}

\item{envelope}{A vector indicating the set of sites that conform the reference envelope (other sites will be compared to the envelope)}

\item{m}{Fuzziness exponent for quality value assessment}

\item{distances_to_envelope}{Flag to indicate that distances to envelope should be included in the result}
}
\value{
\itemize{
   \item{Functions \code{stateEnvelopeVariability} and \code{trajectoryEnvelopeVariability} are used to assess the}
   \item{Functions \code{compareToStateEnvelope} and \code{compareToTrajectoryEnvelope} return data frame with columns identifying the envelope and the Q statistic for the ecological quality with respect to the 
envelope. If \code{nboot.ci != NULL} extra columns are added to indicate the boundaries of a confidence interval for Q, built 
using bootstrap samples of the reference envelope.}
 }
}
\description{
Functions to assess the variability of ecological reference envelopes and to assess the ecological quality of
target stations/observations with respect to reference envelopes.
}
\details{
Functions \code{stateEnvelopeVariability} and \code{trajectoryEnvelopeVariability} are used to assess the 
variability of reference envelopes. Functions \code{compareToStateEnvelope} and \code{compareToTrajectoryEnvelope} are 
used to evaluate the ecological quality of stations/observations with respect to a predefined reference envelope.
}
\examples{
data(glomel)
 
# Extract compositional data matrix
glomel_comp <- as.matrix(glomel[,!(names(glomel) \%in\% c("ID", "Ref", "Complementary"))])
rownames(glomel_comp) <- glomel$ID
 
# Calculate Bray-Curtis distance matrix 
glomel_bc <- vegan::vegdist(glomel_comp, method = "bray")
 
# Define reference envelope (5 stations) by observation ID
glomel_env <- glomel$ID[glomel$Ref]
 
# Assess quality with respect to reference envelope
compareToStateEnvelope(glomel_bc, glomel_env)
 
}
\seealso{
\code{\link{trajectorymetrics}}
}
\author{
Miquel De \enc{Cáceres}{Caceres}, CREAF

Anthony Sturbois, Vivarmor nature, Réserve Naturelle nationale de la Baie de Saint-Brieuc
}
