% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineTrajectories.R
\name{trajectories}
\alias{trajectories}
\alias{defineTrajectories}
\alias{subsetTrajectories}
\title{Trajectory data}
\usage{
defineTrajectories(d, sites, surveys = NULL, times = NULL)

subsetTrajectories(x, site_selection, survey_selection = NULL)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states..}

\item{sites}{A character vector indicating the site corresponding to each ecosystem state (other types are converted to character).}

\item{surveys}{An integer vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{times}{A numeric vector indicating survey times (if missing, survey times are made equal to surveys).}

\item{x}{An object of class \code{trajectories}}

\item{site_selection}{A character vector indicating the subset of site trajectories to be selected.  A different interpretation is used for cycles or fixed date trajectories (see details).}

\item{survey_selection}{An integer vector indicating the subset of surveys to be included (if NULL, all surveys are included).}
}
\value{
An object (list) of class \code{trajectories} with the following elements:
\itemize{
\item{\code{d}: An object of class \code{\link{dist}} containing relationships between ecosystem states}
\item{\code{metadata}: A dataframe describing trajectory states, with the following columns:
\itemize{
\item{\code{sites}: A character vector indicating the site corresponding to each ecosystem state.}
\item{\code{surveys}: An integer vector indicating the survey corresponding to each ecosystem state.}
\item{\code{times}: A numeric vector indicating survey times.}
}
}
}
}
\description{
Defines or subsets data structures for trajectory analysis
}
\details{
When using function \code{subsetTrajectories} on cycles or fixed-date trajectories then the subsetting operation
is assumed to operate on those geometry elements. Hence, the parameter \code{site_selection} applies to cycles or fixed-date trajectories, respectively.
}
\examples{
#Description of sites and surveys
sites <- c("1","1","1","2","2","2")
surveys <- c(1,2,3,1,2,3)
  
#Raw data table
xy<-matrix(0, nrow=6, ncol=2)
xy[2,2]<-1
xy[3,2]<-2
xy[4:6,1] <- 0.5
xy[4:6,2] <- xy[1:3,2]
xy[6,1]<-1

d <- dist(xy)

# Defines trajectories
x <- defineTrajectories(d, sites, surveys)
x

# Extracts (subset) second trajectory
x_2 <- subsetTrajectories(x, "2")
x_2
}
