% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryConvergencePlot.R
\encoding{UTF-8}
\name{trajectoryCyclicalPlots}
\alias{trajectoryCyclicalPlots}
\alias{trajectoryConvergencePlot}
\alias{cycleShiftsArrows}
\title{Summary plot for trajectory convergence and divergence}
\usage{
trajectoryConvergencePlot(
  x,
  type = "pairwise.asymmetric",
  alpha.filter = NULL,
  traj.colors = "grey",
  traj.names = NULL,
  traj.names.colors = "black",
  ...,
  radius = 1,
  conv.color = "red",
  div.color = "blue",
  half.arrows.size = 1,
  tau.links.transp = 0.3,
  top = "between",
  pointy = FALSE
)

cycleShiftsArrows(
  cycle.shifts,
  radius = 1,
  top = "between",
  cycle.shifts.inf.conf = NULL,
  cycle.shifts.sup.conf = NULL,
  arrows.length.mult = "auto",
  arrows.lwd = 2
)
}
\arguments{
\item{x}{An object of class \code{\link{trajectories}}.}

\item{type}{A string indicating the convergence test, either "pairwise.asymmetric", "pairwise.symmetric" or "both" (see \code{\link{trajectoryConvergence}}).}

\item{alpha.filter}{The minimum p-value for a link to be drawn (see \code{\link{trajectoryConvergence}}). Defaults to NULL (all links drawn).}

\item{traj.colors}{The colors for the trajectories (circles). Defaults to "grey".}

\item{traj.names}{The names of trajectories. Defaults to the names provided in \code{x}.}

\item{traj.names.colors}{The color of the names of trajectories on the circles. Defaults to "black".}

\item{...}{Additional parameters passed to \code{\link{polygon}} to personalize the circles representing trajectories.}

\item{radius}{The radius of the circles representing trajectories. Defaults to 1.}

\item{conv.color}{The color used to mark convergent trajectories. Defaults to "red".}

\item{div.color}{The color used to mark divergent trajectories. Defaults to "blue".}

\item{half.arrows.size}{A multiplication coefficient for the size of the arrow heads when representing asymmetric tests results. Defaults to 1.}

\item{tau.links.transp}{The transparency of the links representing the tau statistic of the Mann.Kendall test (see \code{\link{trajectoryConvergence}}).}

\item{top}{A string indicating whether the top of the plotting area should contain a circle representing a trajectory ("top") or should be in between two circles ("between"). Defaults to "between".}

\item{pointy}{Boolean. Should the circles representing trajectories be made pointy? Useful in the context of CETA to represent fixed date trajectories (see \code{\link{trajectoryCyclical}}).}

\item{cycle.shifts}{Cyclical shifts computed for each fixed date trajectory plotted by \code{trajectoryConvergencePlot}.}

\item{cycle.shifts.inf.conf}{Lower confidence intervals for cyclical shifts.}

\item{cycle.shifts.sup.conf}{Upper confidence intervals for cyclical shifts.}

\item{arrows.length.mult}{A multiplication coefficient for the arrows representing cyclical shifts. Attempts an automatic adjustment by default (dividing by max(cycle.shifts)).}

\item{arrows.lwd}{Line width of the arrows. Defaults to 2.}
}
\description{
Provides plots to represent trajectory convergence and divergence tests performed by the function \code{\link{trajectoryConvergence}}. Some CETA-septic options are also available to study changes in cycles shapes.
}
\details{
The function \code{trajectoryConvergencePlot} provides ways to visualize pairwise convergence and divergence between trajectories using calls to function \code{\link{trajectoryConvergence}} which performs the tests.
In the plots, trajectories are represented by circles. The convergence or divergence between pairs of trajectories are represented by links. If convergence tests are symmetric, the links are simple. If the convergence tests are asymmetric, the links are displayed as half arrows pointing from the trajectory converging or diverging towards the trajectory being approached or diverged from.
The width and color hue of the links are proportional to the tau statistic of the Mann.Kendall test performed by the \code{\link{trajectoryConvergence}} function.
The function \code{trajectoryConvergencePlot} also offers the possibility to plot both tests at the same time.

Cyclical Ecological Trajectory Analysis (CETA) specific options are particularly designed to study fixed-date trajectories convergence/divergence patterns:
\itemize{
\item{First, setting \code{pointy = TRUE} when the studied trajectories are the outputs of \code{\link{extractFixedDateTrajectories}} allows to suggest cyclicity.}
\item{Second, the companion function \code{cycleShiftsArrows} allows to represent cyclical shifts by adding arrows to the circles representing trajectories. Clockwise arrows will represent advances, anticlockwise arrows will represent delays. Arrows length is proportional to the cyclical shift provided.
Relevant measures of cyclical shifts have to be computed by the user. A variety of methods may be employed but the outputs of \code{\link{cycleShifts}} cannot be used immediately as they do not directly correspond to the fixed-date trajectories. An example of how to compute relevant measures of cyclical shifts is provided in the CETA vignette.
The arguments \code{radius} and \code{top} in \code{cycleShiftsArrows} must match those in the corresponding \code{trajectoryConvergencePlot} for proper display.}
}
}
\examples{
data("avoca")
avoca_D_man <- vegclust::vegdiststruct(avoca_strat, 
                                       method="manhattan", 
                                       transform = function(x){log(x+1)})
years <- c(1971, 1974, 1978, 1983, 1987, 1993, 1999, 2004, 2009)
avoca_times <- years[avoca_surveys]
avoca_x <- defineTrajectories(d = avoca_D_man,  
                              sites = avoca_sites, 
                              times = avoca_times)

#Raw output with asymmetric convergence test (default)
trajectoryConvergencePlot(avoca_x)

#More refined output with both type of tests and only plotting significant test results (p-value < 0.05)
trajectoryConvergencePlot(avoca_x,type="both",alpha.filter=0.05)

#Much more refined output with nicer colors bigger half arrows and personalized trajectory names
trajectoryConvergencePlot(avoca_x,type="both",alpha.filter=0.05,
                          #adjusting the attributes of the links between trajectories
                          half.arrows.size = 1.5,conv.color = "orangered",div.color="dodgerblue",
                          #controling the size of circles representing trajectories (links are enlarged to match it)
                          radius=1.2, 
                          #customizing the border of the circles representing trajectories
                          traj.colors = "black",border="white",lwd=2,
                          #giving trajectories new names and adjusting the names color
                          traj.names=LETTERS[1:8],traj.names.colors="white")


#CETA-specific example with made-up cyclical shifts:
trajectoryConvergencePlot(avoca_x,type="both",alpha.filter = 0.05,half.arrows.size = 2,pointy=T)

CS <- c(2.5,1.5,0.5,0.5,3,1,1,2)#the made up cyclical shifts (matching the number of trajectories in avoca_x)
cycleShiftsArrows(CS,cycle.shifts.inf.conf = CS-0.2,cycle.shifts.sup.conf = CS+0.2)


}
\seealso{
\code{\link{trajectoryConvergence}},\code{\link{cycleShifts}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
