% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryCyclicalPlots.R
\encoding{UTF-8}
\name{trajectoryCyclicalPlots}
\alias{trajectoryCyclicalPlots}
\alias{cyclePCoA}
\alias{fixedDateTrajectoryPCoA}
\alias{CETA}
\alias{plotting}
\alias{functions:}
\alias{\itemize{}
\alias{\item{Function}
\alias{\code{cyclePCoA}}
\alias{performs}
\alias{principal}
\alias{coordinates}
\alias{analysis}
\alias{(\code{\link{cmdscale}})}
\alias{and}
\alias{draws}
\alias{cycles}
\alias{in}
\alias{the}
\alias{ordination}
\alias{scatterplot.}
\alias{Sister}
\alias{function}
\alias{of}
\alias{\code{trajectoryPCoA}}
\alias{adapted}
\alias{to}
\alias{cycles.}}
\alias{\code{fixedDateTrajectoryPCoA}}
\alias{fixed-date}
\alias{trajectories}
\alias{fixed}
\alias{date-trajectories.}}
\alias{}}
\title{Cyclical trajectory plots}
\usage{
cyclePCoA(
  inputCyclePCoA,
  centered = FALSE,
  sites.colors = NULL,
  print.names = FALSE,
  axes = c(1, 2),
  ...
)

fixedDateTrajectoryPCoA(
  inputFixedDateTrajectoryPCoA,
  sites.lty = NULL,
  print.names = FALSE,
  axes = c(1, 2),
  ...
)
}
\arguments{
\item{inputCyclePCoA}{The full output of function \code{\link{extractCycles}}.}

\item{centered}{Boolean. Have the cycles been centered? Default to FALSE.}

\item{sites.colors}{The colors applied to the different sites. The cycles will be distinguished (old to recent) by increasingly lighter tones of the provided colors.}

\item{print.names}{A boolean flag to indicate whether the names of cycles or fixed-date trajectories should be printed.}

\item{axes}{The pair of principal coordinates to be plotted.}

\item{...}{Additional parameters for function \code{\link{arrows}}.}

\item{inputFixedDateTrajectoryPCoA}{The full output of function \code{\link{extractFixedDateTrajectories}}.}

\item{sites.lty}{The line type for the different sites (see \code{\link{par}}, \code{"lty"}). The fixed-date trajectories will be distinguished by a default circular color palette.}
}
\value{
Functions \code{cyclePCoA} and \code{fixedDateTrajectoryPCoA} return the results of calling of \code{\link{cmdscale}}.
}
\description{
Set of plotting functions for Cyclical Ecological Trajectory Analysis:
}
\details{
\itemize{
\item{Function \code{cyclePCoA} performs principal coordinates analysis (\code{\link{cmdscale}}) and draws trajectories in the ordination scatterplot.}
\item{Function \code{fixedDateTrajectoryPCoA} performs principal coordinates analysis (\code{\link{cmdscale}}) and draws trajectories in the ordination scatterplot.}
}

The functions \code{cyclePCoA} and \code{fixedDateTrajectoryPCoA} give adapted graphical representation of cycles and fixed-date trajectories respectively using principal coordinate analysis (PCoA, see \code{\link{cmdscale}}).
Function \code{cyclePCoA} handles external and potential interpolated ecological states so that they are correctly taken in account in PCoA (i.e. avoiding duplication, and reducing the influence of interpolated ecological states as much as possible). In case of centered cycles, the influence of these point will grow as they will not correspond to duplications anymore.
In case of centered cycles, the intended use is to set the parameter \code{centered} to \code{TRUE}.
Function \code{fixedDateTrajectoryPCoA} has no flexibility for the colors as it defaults to a circular color palette adapted to representation of cyclical processes.
}
\seealso{
\code{\link{trajectoryCyclical}}, \code{\link{cmdscale}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
