% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryCyclicalPlots.R
\encoding{UTF-8}
\name{trajectoryCyclicalPlots}
\alias{trajectoryCyclicalPlots}
\alias{cyclePCoA}
\alias{fixedDateTrajectoryPCoA}
\alias{customCircularPalette}
\title{Cyclical trajectory plots}
\usage{
cyclePCoA(
  x,
  centered = FALSE,
  sites.colors = NULL,
  print.names = FALSE,
  axes = c(1, 2),
  ...
)

fixedDateTrajectoryPCoA(
  x,
  fixedDates.colors = NULL,
  sites.lty = NULL,
  print.names = FALSE,
  axes = c(1, 2),
  ...
)
}
\arguments{
\item{x}{The full output of function \code{\link{extractCycles}} or \code{\link{extractFixedDateTrajectories}} as appropriate, an object of class \code{\link{trajectories}}.}

\item{centered}{Boolean. Have the cycles been centered? Default to FALSE.}

\item{sites.colors}{The colors applied to the different sites. The cycles will be distinguished (old to recent) by increasingly lighter tones of the provided colors.}

\item{print.names}{A boolean flag to indicate whether the names of cycles or fixed-date trajectories should be printed.}

\item{axes}{The pair of principal coordinates to be plotted.}

\item{...}{Additional parameters for function \code{\link{arrows}}.}

\item{fixedDates.colors}{The colors applied to the different fixed dates trajectories. Defaults to a simple RGB circular color palette.}

\item{sites.lty}{The line type for the different sites (see \code{\link{par}}, \code{"lty"}).}
}
\value{
Functions \code{cyclePCoA} and \code{fixedDateTrajectoryPCoA} return the results of calling of \code{\link{cmdscale}}.
}
\description{
Plotting functions for Cyclical Ecological Trajectory Analysis:
}
\details{
\itemize{
\item{Function \code{cyclePCoA} removes unwanted points (see details) and performs principal coordinates analysis (\code{\link{cmdscale}}) and draws cycles in the ordination scatterplot.}
\item{Function \code{fixedDateTrajectoryPCoA} performs principal coordinates analysis (\code{\link{cmdscale}}) and draws fixed date trajectories in the ordination scatterplot.}
}

The functions \code{cyclePCoA} and \code{fixedDateTrajectoryPCoA} give adapted graphical representation of cycles and fixed-date trajectories respectively using principal coordinate analysis (PCoA, see \code{\link{cmdscale}}).
Function \code{cyclePCoA} handles external and potential interpolated ecological states so that they are correctly taken in account in PCoA (i.e. avoiding duplication, and reducing the influence of interpolated ecological states as much as possible). In case of centered cycles, the influence of these ecological states will grow as they will not correspond to duplications anymore.
In case of centered cycles, the intended use is to set the parameter \code{centered} to \code{TRUE}.
}
\seealso{
\code{\link{trajectoryCyclical}}, \code{\link{cmdscale}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
