% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryRMA.R
\name{trajectoryRMA}
\alias{trajectoryRMA}
\title{Relative Trajectory Movement Assessment (RTMA)}
\usage{
trajectoryRMA(x, alpha = 0.05, nperm = 999, full.output = TRUE, add = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{trajectories}}.}

\item{alpha}{The alpha level for the tests performed in RTMA. Defaults to \code{0.05}.}

\item{nperm}{Passed to function \code{\link{trajectoryCorrespondence}}. The number of permutations to be used in the dynamic correspondence test. Defaults to \code{999}.}

\item{full.output}{Flag to indicate that the full output of tests should be computed. Defaults to \code{TRUE}. Setting to FALSE will improve computation speed but yield incomplete outputs (see details).}

\item{add}{Passed to function \code{\link{trajectoryConvergence}}. Flag to indicate that constant values should be added (local transformation) to correct triplets of distance values that do not fulfill the triangle inequality.}
}
\value{
Function \code{trajectoryRMA} returns an object of class \code{\link{list}} containing:
\itemize{
\item{\code{relationship}, a matrix containing the relative movements relationships attributed to each pair of trajectories.}
\item{\code{symmetricConvergence}, a list containing the results of the symmetric convergence test.}
\item{\code{asymmetricConvergence}, a list containing the results of the two asymmetric convergence tests.}
\item{\code{dynamicCorrespondence}, a matrix containing the results of the the dynamic correspondence tests (partial if \code{full.out = FALSE}).}
\item{\code{parameters}, a vector containing the parameters \code{alpha}, the \enc{Šidák}{Sidak} corrected \code{alpha}, and \code{nperm}}.
}
In addition to the relationships recognized by RTMA, matrix \code{relationship} provides details on asymmetric relationships (namely \code{"Approaching"}, \code{"Approaching-Stationary"}, \code{"Departing"}, \code{"Departing-Stationary"}, \code{"Pursuit"},  \code{"Catch-up"},  \code{"Escape"}).
In asymmetric relationships, the two trajectories have different behavior denoted by descriptive suffixes pasted on the relationship labels using \code{"_"} as a separator (e.g. \code{"Departing_Departer"}). In the matrix \code{relationship}, the suffixes apply to the ROW trajectories.
}
\description{
Relative Trajectory Movement Assessment (RTMA) is a method allowing testing and qualifying of the relative movements of ecological trajectories (e.g. "convergence", "parallel" etc., see details) as described in Djeghri et al. (in prep). It is implemented in function \code{trajectoryRMA()}.
}
\details{
Function \code{trajectoryRMA} attributes a relationship to pairs of ecological trajectories A and B describing their relative movement. It does so by combining four tests:
\itemize{
\item{Three convergence tests performed through internal callings of function \code{\link{trajectoryConvergence}}:
\itemize{
\item{The symmetric convergence test between trajectories A and B.}
\item{The asymmetric convergence test assessing if trajectory A approaches trajectory B.}
\item{The asymmetric convergence test assessing if trajectory B approaches trajectory A.}
}
}
\item{One dynamic correspondence test performed through internal callings to function \code{\link{trajectoryCorrespondence}}.}
}
To account for multiple testing, \code{trajectoryRMA} performs internally a \enc{Šidák}{Sidak} (1967) correction on the alpha level provided in parameter \code{alpha}.

The results of the four tests (p-values and sign of statistic) are used to assign to each trajectory pair a relationship describing their relative movements. RTMA recognizes a total of 12 relationships, some existing in "weak" variations:
\itemize{
\item{\code{"Convergence"} relationship: The two trajectories converge. Exists in a weak version.}
\item{\code{"Divergence"} relationship: The two trajectories diverge. Exists in a weak version.}
\item{\code{"Approaching"} relationship: One trajectory approaches the other.}
\item{\code{"Approaching-Stationary"} relationship: As \code{"Approaching"} but the trajectory approached is stationary relative to the approaching trajectory.}
\item{\code{"Departing"} relationship: One trajectory moves away from the other.}
\item{\code{"Departing-Stationary"} relationship: As \code{"Departing"} but the trajectory departed from is stationary relative to the departing trajectory.}
\item{\code{"Pursuit"} relationship: The two trajectories follow each other.}
\item{\code{"Catch-up"} relationship: As \code{"Pursuit"} but the following trajectory moves faster.}
\item{\code{"Escape"} relationship: As \code{"Pursuit"} but the leading trajectory is faster.}
\item{\code{"Parallel"} relationship: The two trajectories travel side by side with broadly similar movements.}
\item{\code{"Antiparallel"} relationship: As \code{"Parallel"} but the two trajectories travel in opposite directions.}
\item{\code{"Neutral"} relationship: The two trajectories have no particular movements relative to each other (effectively the Null Hypothesis for RTMA).}
}
Some relationships are asymmetric (e.g. in \code{"Pursuit"} there is a leading and a following trajectory).
In these asymmetric relationships the output of function \code{trajectoryRMA} gives the details (see Value section).
In rare cases, unlikely relationships (labelled \code{"Other"}) may occur. These involve contradictory patterns hard to interpret.

LIMITATIONS: RTMA has some limitations, in particular it uses trend tests not well suited to study trajectories pairs with changing relative movements (e.g. if two trajectories cross each other, they are first converging then diverging).
We advise users to not only rely on RTMA but to also visualize trajectories using function \code{\link{trajectoryPCoA}} for ecological interpretations. See Djeghri et al. (in prep) for more details.
Note also that, because RTMA incorporates a correction for multiple testing, it needs somewhat long trajectories to operate (minimum number of survey = 6 at alpha = 0.05).

COMPUTATION TIME: The dynamic correspondence tests performed in RTMA are computationally costly permutation tests only used when all three convergence tests are non-significant.
Function \code{trajectoryRMA} performs by default all tests but it is possible to only perform the tests useful for RTMA by setting \code{full.output = FALSE}.
This will reduce computation time but the details of the output of RTMA will not contain the information on all possible dynamic correspondence tests, only on relevant ones.

PLOTTING: Function \code{\link{trajectoryConvergencePlot}} provides options to plot the results of RTMA.
}
\examples{
#Obtain and format some trajectories
data("avoca")
avoca_D_man <- vegclust::vegdiststruct(avoca_strat, 
                                       method ="manhattan", 
                                       transform = function(x){log(x+1)})
years <- c(1971, 1974, 1978, 1983, 1987, 1993, 1999, 2004, 2009)
avoca_times <- years[avoca_surveys]
avoca_x <- defineTrajectories(d = avoca_D_man,  
                              sites = avoca_sites, 
                              times = avoca_times)
                              
#Visualize the trajectories
trajectoryPCoA(avoca_x,traj.colors = RColorBrewer::brewer.pal(8,"Accent"),length=0.1,lwd=2)
legend("bottomleft",bty="n",legend=1:8,col=RColorBrewer::brewer.pal(8,"Accent"),lwd=2,ncol=2)

#Perform RTMA
trajectoryRMA(avoca_x)


}
\references{
Djeghri et al. (in preparation) Uncovering the relative movements of ecological trajectories.

\enc{Šidák}{Sidak}, Z. (1967) Rectangular confidence regions for the means of multivariate normal distributions. Journal of the American Statistical Association 62:648-633.
}
\seealso{
\code{\link{trajectoryConvergence}}, \code{\link{trajectoryCorrespondence}}, \code{\link{trajectoryConvergencePlot}}
}
\author{
Nicolas Djeghri, UBO

Miquel De \enc{Cáceres}{Caceres}, CREAF
}
