% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryUtils.R
\encoding{UTF-8}
\name{trajectoryutils}
\alias{trajectoryutils}
\alias{extractTrajectories}
\alias{centerTrajectories}
\alias{smoothTrajectories}
\alias{is.metric}
\title{Utility functions for Ecological Trajectory Analysis}
\usage{
extractTrajectories(d, sites, selection, surveys = NULL)

smoothTrajectories(
  d,
  sites,
  surveys = NULL,
  survey_times = NULL,
  kernel_scale = 1,
  fixed_endpoints = TRUE
)

centerTrajectories(d, sites, exclude = integer(0), verbose = FALSE)

is.metric(d, tol = 1e-04)
}
\arguments{
\item{d}{A symmetric \code{\link{matrix}} or an object of class \code{\link{dist}} containing the distance values between pairs of ecosystem states (see details).}

\item{sites}{A vector indicating the site corresponding to each ecosystem state.}

\item{selection}{A character vector of sites, a numeric vector of site indices or logical vector of the same length as \code{sites}, indicating a subset of site trajectories to be selected.}

\item{surveys}{A vector indicating the survey corresponding to each ecosystem state (only necessary when surveys are not in order).}

\item{survey_times}{A vector indicating the survey time for all surveys (if \code{NULL}, time between consecutive surveys is considered to be one)}

\item{kernel_scale}{Scale of the Gaussian kernel, related to survey times}

\item{fixed_endpoints}{A logical flag to force keeping the location of trajectory endpoints unmodified}

\item{exclude}{An integer vector indicating sites that are excluded from trajectory centroid computation}

\item{verbose}{Provides console output informing about process (useful for large dataset).}

\item{tol}{Tolerance value for metricity}
}
\value{
Function \code{extractTrajectories} return a list of two elements:
\itemize{
\item{\code{d}: an object of class \code{\link{dist}}, the new distance matrix describing resemblance between states of the selected trajectories.}
\item{\code{metadata}: an object of class \code{\link{data.frame}} describing the ecological states in \code{d} with columns:
\itemize{
\item{\code{sites}: The site associated to each ecological state.}
\item{\code{surveys}: The survey associated to each ecological state.}
}
}
}
Functions \code{centerTrajectories}, \code{smoothTrajectories} return an object of class \code{\link{dist}}.
}
\description{
The set following set of utility functions are provided:
\itemize{
\item{Function \code{extractTrajectories} allows selecting a subset of trajectories corresponding to specific sites.}
\item{Function \code{smoothTrajectories} performs multivariate smoothing on trajectory data using a Gaussian kernel.}
\item{Function \code{centerTrajectories} shifts all trajectories to the center of the multivariate space and returns a modified distance matrix.}
\item{Function \code{is.metric} checks whether the input dissimilarity matrix is metric (i.e. all triplets fulfill the triangle inequality).}
}
}
\details{
Details of calculations are given in De \enc{Cáceres}{Caceres} et al (2019).
Function \code{centerTrajectories} performs centering of trajectories using matrix algebra as explained in Anderson (2017).
}
\references{
De \enc{Cáceres}{Caceres} M, Coll L, Legendre P, Allen RB, Wiser SK, Fortin MJ, Condit R & Hubbell S. (2019). Trajectory analysis in community ecology. Ecological Monographs 89, e01350.

Anderson (2017). Permutational Multivariate Analysis of Variance (PERMANOVA). Wiley StatsRef: Statistics Reference Online. 1-15. Article ID: stat07841.
}
\seealso{
\code{\link{trajectoryplots}} \code{\link{trajectorymetrics}}
}
\author{
Miquel De \enc{Cáceres}{Caceres}, CREAF

Nicolas Djeghri, UBO
}
